backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.1.0
inference-sdk==0.9.17
requests==2.31.0
python-multipart==0.0.6
gunicorn==21.2.0
